move read functions from option utils to fix cyclic import
